addressed comments by Alexey
