+++ connectors/katalog/utils/utils.go
@@ -4,9 +4,52 @@
 package utils
 
 import (
+	"context"
 	"crypto/rand"
+	"fmt"
 	"math/big"
+	"sort"
+	"strings"
+
+	"fybrik.io/fybrik/connectors/katalog/pkg/apis/katalog/v1alpha1"
+	"github.com/rs/zerolog"
+	kclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
+
+type sortByLength []string
+
+// Len implements Len of sort.Interface
+func (s sortByLength) Len() int {
+	return len(s)
+}
+
+// Swap implements Swap of sort.Interface
+func (s sortByLength) Swap(i, j int) {
+	s[i], s[j] = s[j], s[i]
+}
+
+// Less implements Less of sort.Interface
+func (s sortByLength) Less(i, j int) bool {
+	return len(s[i]) > len(s[j])
+}
+
+func getLongest(toFind []string) []string {
+	// We sort it by length, descending
+	sort.Sort(sortByLength(toFind))
+	longest := []string{toFind[0]}
+
+	// In case we have more than one element in toFind...
+	if len(toFind) > 1 {
+		for _, str := range toFind[1:] {
+			if len(str) < len(longest[0]) {
+				break
+			}
+			longest = append(longest, str)
+		}
+	}
+	fmt.Println(longest)
+	return longest
+}
 
 // GenerateRandomString returns a securely generated random string.
 // It will return an error if the system's secure random
@@ -41,4 +84,41 @@ func GenerateRandomNumber(min, max int64) int64 {
 
 	// add n to min to support the passed in range
 	return n.Int64() + min
+}
+
+func GenerateUniqueAssetName(namespace, namePrefix string, log zerolog.Logger, client kclient.Client) (string, error) {
+	var result v1alpha1.AssetList
+	var randomStringLength = 1
+	var uniqueAssetName = ""
+	err := client.List(context.Background(), &result, kclient.InNamespace(namespace))
+	if err == nil {
+		listOfCandidates := make([]string, 0)
+		for i := 0; i < len(result.Items); i++ {
+			if strings.Contains(result.Items[i].Spec.Metadata.Name, namePrefix) {
+				listOfCandidates = append(listOfCandidates, result.Items[i].Spec.Metadata.Name)
+			}
+		}
+		log.Info().Msg("listOfCandidates : " + strings.Join(listOfCandidates, "|"))
+		var randomStr string
+		randomStr, err = GenerateRandomString(randomStringLength)
+		if err == nil {
+			if len(listOfCandidates) > 0 {
+				longestArr := getLongest(listOfCandidates)
+				const delimiter = "|"
+				log.Info().Msg("longestArr : " + strings.Join(longestArr, delimiter))
+				randIdx := GenerateRandomNumber(0, int64(len(longestArr)))
+				log.Info().Msg("randIdx : " + fmt.Sprint(randIdx))
+				uniqueAssetName = longestArr[randIdx] + randomStr
+			} else {
+				// no asset with the given prefix
+				uniqueAssetName = namePrefix + randomStr
+			}
+			log.Info().Msg("uniqueAssetName generated : " + uniqueAssetName)
+		} else {
+			log.Info().Msg("Error during GenerateRandomString: " + err.Error())
+		}
+	} else {
+		log.Info().Msg("Error during list operation: " + err.Error())
+	}
+	return uniqueAssetName, err
 }
